lang/rust: Move helloworld in its own directory
Move Rust helloworld implementation and support files in a dedicated
directory (`lang/rust/helloworld/`). This is done to allow other Rust
sample programs to be added in a separate directory. That directory will
be used as the root filesystem in the bincompat run.

Everything could still be in one directory, and use that as the root
filesystem. But it would fill a lot of space when packing the filesystem
in a CPIO archive and mounting it as an initial ramdisk image.

Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>