lang/rust: Introduce simple HTTP server
Introduce a simple HTTP server written in Rust. It listens for
connections on port 8080 and server a helloworld HTML message.

As with other bincompat applications, add both input and generated
files. Input files are `http_server.rs` and `Makefile`. Generated files
are the `http_server` dynamic executable and the shared libraries it
used (in `lib/` and `lib64/`).

Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>